chore: rename project to FLAME Hub Python Client
